Updates with latest version of code
This version performs slower than previous versions because it is a more strict implementation of TD(0) algorithm. I think the previous version would assign some of the values of terminal states to the rewards immediately, rather than the algorithm having to 'learn' that the terminal state's value was consistently equal to the actual reward.